# docs/edgelet/edgelet-api-v1-openapi.yaml
openapi: 3.1.0
info:
  title: Edgelet API
  version: 1.0.0
  description: >
    Edgelet API v1 baseline for Edgelet daemon and CLI.
    Canonical namespace is /v1. Any path/schema/auth deviation must be
    explicitly approved before implementation.
servers:
  - url: https://edgelet.default.svc.bridge.local
    description: HTTPS/WSS transport for microservices and local clients
  - url: http+unix://%2Frun%2Fedgelet%2Fedgelet.sock
    description: Unix socket transport for CLI/admin operations
security:
  - BearerAuth: []
paths:
...
  /v1/system/controlplane:
    get:
      tags: [System]
      summary: Get local ControlPlane deployment status
      description: Admin-only. Returns singleton controller deployment status.
      responses:
        "200":
          description: ControlPlane status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlPlaneStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [System]
      summary: Delete local ControlPlane deployment
      description: Admin-only. Stops the controller container, removes volumes, and deletes the SQLite row.
      responses:
        "200":
          description: ControlPlane deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/system/controlplane/manifest:
    get:
      tags: [System]
      summary: Get ControlPlane manifest with secrets masked
      responses:
        "200":
          description: Masked manifest YAML
          content:
            application/json:
              schema:
                type: object
                properties:
                  manifestYaml:
                    type: string
                  masked:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/deploy/controlplane:apply:
    post:
      tags: [Deploy]
      summary: Apply ControlPlane manifest (async)
      description: Admin-only. Accepts the manifest and runs apply in the background. Poll GET /v1/deploy/controlplane:apply/{operationId}. At most one apply may be running; a second POST returns 409 APPLY_IN_PROGRESS.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DeployManifestMultipartRequest"
      responses:
        "202":
          description: Apply accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlPlaneApplyOperationResponse"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /v1/deploy/controlplane:validate:
    post:
      tags: [Deploy]
      summary: Validate ControlPlane manifest
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DeployManifestMultipartRequest"
      responses:
        "200":
          description: Validation result
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    InvalidArgument:
      description: Invalid argument
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    ApiError:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            requestId:
              type: string
...
    ControlPlaneStatus:
      type: object
      required: [controllerUuid, namespace, name, runtimeState, source, type]
      properties:
        controllerUuid:
          type: string
        namespace:
          type: string
        name:
          type: string
        image:
          type: string
        containerId:
          type: string
        state:
          type: string
        desiredState:
          type: string
        runtimeState:
          type: string
        lastError:
          type: string
        restartCount:
          type: integer
        generation:
          type: integer
        observedGeneration:
          type: integer
        lastTransitionAt:
          type: integer
          format: int64
        source:
          type: string
          enum: [controlplane]
        type:
          type: string
          enum: [controlplane]
...
    DeployManifestMultipartRequest:
      type: object
      required: [manifest]
      properties:
        manifest:
          type: string
          format: binary
          description: YAML document; microservice deploys use LocalDeployMicroserviceManifest (spec.image)
        sourceName:
          type: string
        dryRun:
          type: boolean
        async:
          type: string
          description: Optional boolean-like field used by deploy handlers that support operation polling.
...